Fix signup filtering by approval (#105)
On the request credit page, the filtering to only include non-approved requests was not effective, so we would query all signups. 
On top of that, we were also sorting data chronologically, so we were only getting the 50 oldest approvals.

For folks with more than 50 signups, we would therefore not load the more recent office hours, and desk workers couldn't request credit for them.